Normalize Olsson height maps
Preserve the source-compatible raw integer golden boundary while exposing normalized float64 elevations through the package and JSON adapter. Define flat maps as all-zero output and test JSON round trips.

Fixes #11

Amp-Thread-ID: https://ampcode.com/threads/T-01a0d5e2-6951-74d4-b028-329904f1aebb

+++ cmd/olsson/main.go
@@ -28,11 +28,11 @@ import (
 )
 
 type output struct {
-	Seed       int64 `json:"seed"`
-	Width      int   `json:"width"`
-	Height     int   `json:"height"`
-	Faults     int   `json:"faults"`
-	Elevations []int `json:"elevations"`
+	Seed       int64     `json:"seed"`
+	Width      int       `json:"width"`
+	Height     int       `json:"height"`
+	Faults     int       `json:"faults"`
+	Elevations []float64 `json:"elevations"`
 }
 
 func main() {
